refactor(dal): split src/lib/dal.ts by surface (#278)
Splits the 1539-line `src/lib/dal.ts` chokepoint into per-surface modules
under `src/lib/dal/` with an `index.ts` barrel that preserves the
existing `@/lib/dal` import paths. Pure organization — no query, RLS, or
behavior changes.

Layout:
- `types.ts`         shared scoping types + `getVisibleDeviceIds` (149)
- `user.ts`          auth + org members (40)
- `overview.ts`      stats / daily activity / heatmap / cost-by-user (358)
- `team.ts`          team activity series (63)
- `devices.ts`       device cost + activity (192)
- `models.ts`        model cost + activity (114)
- `repos.ts`         cost by repo / branch / ticket (146)
- `surfaces.ts`      surface filter + spend (95)
- `sessions.ts`      sessions list + detail (277)
- `sync.ts`          freshness snapshot (98)
- `pricing.ts`       price-list + recalc runs (89)
- `index.ts`         barrel re-export (64)

`dal.test.ts` (40 tests, dynamic-imported via `@/lib/dal`) passes
unchanged; full suite (380 tests), typecheck, lint, and `next build`
all green.
