--- src/lib/dal.ts
@@ -1,1539 +1,0 @@
-import "server-only";
-import { createClient } from "@/lib/supabase/server";
-import { createAdminClient } from "@/lib/supabase/admin";
-
-export interface DateRange {
-  /** Inclusive lower bound in the **viewer's local TZ** (`YYYY-MM-DD`). */
-  from: string;
-  /** Inclusive upper bound in the **viewer's local TZ** (`YYYY-MM-DD`). */
-  to: string;
-  /**
-   * UTC `bucket_day` lower bound for the daemon's UTC-bucketed
-   * `daily_rollups` table. Derived from `from 00:00:00` in the viewer's TZ
-   * so the SQL filter captures every UTC bucket overlapping the local-TZ
-   * window — including the previous UTC day for users west of UTC, where
-   * yesterday-evening-local activity lands in "today's" UTC bucket. See
-   * siropkin/budi-cloud#78.
-   */
-  bucketFrom: string;
-  /** UTC `bucket_day` upper bound; mirror of `bucketFrom`. */
-  bucketTo: string;
-  /**
-   * Inclusive lower bound for `session_summaries.started_at`, an ISO-8601
-   * UTC instant (e.g. `2026-04-26T07:00:00.000Z`). Sessions are precise
-   * timestamps so we filter on the actual instant rather than a calendar
-   * day, avoiding the same TZ-vs-UTC drift that motivates `bucketFrom`.
-   */
-  startedAtFrom: string;
-  /** Inclusive upper bound for `session_summaries.started_at`. */
-  startedAtTo: string;
-}
-
-interface BudiUser {
-  id: string;
-  org_id: string | null;
-  role: string;
-  api_key: string;
-  display_name: string | null;
-  email: string | null;
-}
-
-/**
- * Optional scoping options for the dashboard breakdown queries.
- *
- * `scopedUserId` narrows the visible-device set to a single teammate's devices
- * — the manager-only header filter introduced in #80. It is silently ignored
- * for member viewers (their visibility is already self-only per ADR-0083 §6)
- * and silently falls back to the org-wide set when the id is unknown or
- * belongs to another org, mirroring the existing role branch in
- * `getVisibleDeviceIds`. We deliberately do not surface a 4xx so an attacker
- * can't enumerate other-org user ids by probing this parameter.
- */
-export interface ScopeOptions {
-  scopedUserId?: string | null;
-  /**
-   * Narrow every aggregation to one or more surfaces (#187), e.g. `['vscode']`
-   * for the JetBrains-vs-VS Code rollout question. `null`, `undefined`, or an
-   * empty array all mean "no filter" — the dashboard's `<SurfaceFilter>` chip
-   * collapses an all-deselected state back to the default rather than zeroing
-   * the page out. The breakdown RPCs treat `NULL p_surfaces` and an empty
-   * array identically (015) so callers don't have to remember which sentinel
-   * to pass.
-   */
-  surfaces?: string[] | null;
-}
-
-/**
- * Normalize a `ScopeOptions.surfaces` value to the wire shape the RPCs want.
- * Returns `null` for any "no filter" case; otherwise a deduped, non-empty
- * array. Centralized so callers can pass a raw URL-derived value without
- * branching on undefined / [] themselves.
- */
-function normalizeSurfaces(raw: string[] | null | undefined): string[] | null {
-  if (!raw || raw.length === 0) return null;
-  const trimmed = raw.map((s) => s.trim()).filter((s) => s.length > 0);
-  if (trimmed.length === 0) return null;
-  return Array.from(new Set(trimmed));
-}
-
-/**
- * Get the current budi user and verify they have an org.
- * Uses admin client because the auth→users mapping needs to bypass RLS
- * during the initial lookup.
- */
-export async function getCurrentUser(): Promise<BudiUser | null> {
-  const supabase = await createClient();
-  const {
-    data: { user: authUser },
-  } = await supabase.auth.getUser();
-  if (!authUser) return null;
-
-  const admin = createAdminClient();
-  const { data } = await admin
-    .from("users")
-    .select("id, org_id, role, api_key, display_name, email")
-    .eq("id", authUser.id)
-    .single();
-
-  return data;
-}
-
-/**
- * Get overview stats visible to the current user.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- *
- * Aggregation runs server-side via `dashboard_overview_stats` (#92) so the
- * sums are independent of any PostgREST row cap — `getOverviewStats` and
- * every breakdown query agree on the same row set regardless of org size.
- *
- * The session count is filtered by the same `bucket_day` calendar range as
- * the rollup totals (#155). Pre-#155 the count used a precise TIMESTAMPTZ
- * window over `started_at`, which excluded sessions with NULL `started_at`
- * and was narrower than the rollup window — on `?days=1` that asymmetry
- * was enough to collapse the previous-period count to zero while every
- * other card on the same row showed real period-over-period deltas.
- */
-export interface OverviewStats {
-  totalCostCents: number;
-  totalCostCentsIngested: number;
-  totalInputTokens: number;
-  totalOutputTokens: number;
-  totalMessages: number;
-  totalSessions: number;
-}
-
-export async function getOverviewStats(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-): Promise<OverviewStats> {
-  const admin = createAdminClient();
-
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) {
-    return {
-      totalCostCents: 0,
-      totalCostCentsIngested: 0,
-      totalInputTokens: 0,
-      totalOutputTokens: 0,
-      totalMessages: 0,
-      totalSessions: 0,
-    };
-  }
-
-  const { data, error } = await admin.rpc("dashboard_overview_stats", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  const row = (data?.[0] ?? null) as OverviewRow | null;
-  // Pre-#235 the RPC didn't surface `_ingested`; default to the effective
-  // total so a deployment that ships the dashboard ahead of migration 020
-  // simply reports zero savings rather than crashing on `undefined`.
-  const effective: number = Number(row?.total_cost_cents ?? 0);
-  const ingested: number = Number(row?.total_cost_cents_ingested ?? effective);
-  return {
-    totalCostCents: effective,
-    totalCostCentsIngested: ingested,
-    totalInputTokens: Number(row?.total_input_tokens ?? 0),
-    totalOutputTokens: Number(row?.total_output_tokens ?? 0),
-    totalMessages: Number(row?.total_messages ?? 0),
-    totalSessions: Number(row?.total_sessions ?? 0),
-  };
-}
-
-interface OverviewRow {
-  total_cost_cents: number | string;
-  total_cost_cents_ingested?: number | string;
-  total_input_tokens: number | string;
-  total_output_tokens: number | string;
-  total_messages: number | string;
-  total_sessions: number | string;
-}
-
-/**
- * Get daily cost activity for charts.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- *
- * Aggregation runs server-side via `dashboard_daily_activity` (#92). Prior to
- * #92 this client-side reduce silently dropped the *oldest* days at the left
- * edge of the chart once the predicate matched > 100,000 rollup rows.
- */
-export async function getDailyActivity(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-) {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_daily_activity", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as DailyActivityRow[])
-    .map((r) => {
-      const effective = Number(r.cost_cents);
-      return {
-        bucket_day: r.bucket_day,
-        input_tokens: Number(r.input_tokens),
-        output_tokens: Number(r.output_tokens),
-        cost_cents: effective,
-        // Pre-#235 RPC has no `_ingested` column — collapse to effective so
-        // the toggle's "hide when ingested == effective everywhere" check
-        // stays correct and we never paint a fake savings delta.
-        cost_cents_ingested: Number(r.cost_cents_ingested ?? effective),
-        message_count: Number(r.message_count),
-      };
-    })
-    .sort((a, b) => a.bucket_day.localeCompare(b.bucket_day));
-}
-
-interface DailyActivityRow {
-  bucket_day: string;
-  input_tokens: number | string;
-  output_tokens: number | string;
-  cost_cents: number | string;
-  cost_cents_ingested?: number | string;
-  message_count: number | string;
-}
-
-/**
- * Day-of-week × hour-of-day session counts for the Overview heatmap (#150).
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- *
- * Buckets are computed in the **viewer's IANA timezone** (server-side via
- * `dashboard_activity_heatmap`) so a US/Pacific viewer's "5pm peak" sits at
- * `hour=17` instead of drifting to UTC's `hour=00`. Falls back to UTC when
- * the viewer's TZ cookie is missing — same fallback `dateRangeFromDays`
- * uses, so the bucketing TZ matches the range-window TZ.
- *
- * Returns 0..168 rows (one per non-empty `(dow, hour)` cell). Empty cells
- * are absent from the result; the client fills them with zero rather than
- * paying for a `generate_series` cross join on every page load.
- */
-export async function getActivityHeatmap(
-  user: BudiUser,
-  range: DateRange,
-  timeZone: string | null,
-  options?: ScopeOptions
-): Promise<HeatmapCell[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_activity_heatmap", {
-    p_device_ids: deviceIds,
-    p_started_from: range.startedAtFrom,
-    p_started_to: range.startedAtTo,
-    p_time_zone: timeZone ?? "UTC",
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as HeatmapRow[]).map((r) => ({
-    dow: Number(r.dow),
-    hour: Number(r.hour),
-    session_count: Number(r.session_count),
-    cost_cents: Number(r.cost_cents),
-  }));
-}
-
-export interface HeatmapCell {
-  dow: number;
-  hour: number;
-  session_count: number;
-  cost_cents: number;
-}
-
-interface HeatmapRow {
-  dow: number | string;
-  hour: number | string;
-  session_count: number | string;
-  cost_cents: number | string;
-}
-
-/**
- * Earliest day (`YYYY-MM-DD`) with a rollup for any device visible to the
- * viewer, or `null` if the org has never synced anything. Used to materialize
- * the `?days=all` sentinel into a concrete `from` before hitting the
- * range-scoped queries so their signatures stay unchanged.
- */
-export async function getEarliestActivity(
-  user: BudiUser,
-  options?: ScopeOptions
-): Promise<string | null> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return null;
-
-  const surfaces = normalizeSurfaces(options?.surfaces);
-  let query = admin
-    .from("daily_rollups")
-    .select("bucket_day")
-    .in("device_id", deviceIds);
-  if (surfaces) query = query.in("surface", surfaces);
-  const { data } = await query
-    .order("bucket_day", { ascending: true })
-    .limit(1)
-    .maybeSingle();
-
-  return (data?.bucket_day as string | null) ?? null;
-}
-
-/** Synthetic user id used to group rollups whose owner we can't surface. */
-export const UNASSIGNED_USER_ID = "__unassigned__";
-
-/**
- * Get cost breakdown by user/device.
- * Manager sees all users; member sees only their own cost (ADR-0083 §6).
- *
- * Rollups that resolve to a visible user are grouped by that user. Any cost
- * left over (devices whose owner isn't in the viewer's visible set, or
- * rollups with no matching device row) is surfaced as an `Unassigned` row,
- * mirroring how `Cost by Project` on the Repos page already handles
- * unattributed data. This guarantees
- *
- *     getOverviewStats.totalCostCents === sum(getCostByUser[...].cost_cents)
- *
- * for the same (user, range), fixing the Overview/Team reconciliation gap
- * described in #15.
- */
-export async function getCostByUser(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-) {
-  const admin = createAdminClient();
-
-  // Use the identical device set as `getOverviewStats` so the two pages
-  // agree on the denominator.
-  const deviceIds = await getVisibleDeviceIds(admin, user);
-  if (deviceIds.length === 0) return [];
-
-  // Aggregate rollups server-side (#92). The owner mapping lives in the small,
-  // bounded `devices` + `users` tables, so the secondary join still happens in
-  // JS without risk of row-cap truncation.
-  const { data: rows, error } = await admin.rpc("dashboard_cost_by_device", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-  const deviceCostRows = (rows ?? []) as DeviceCostRow[];
-
-  const { data: devices } = await admin
-    .from("devices")
-    .select("id, user_id")
-    .in("id", deviceIds);
-
-  const deviceToUser = new Map<string, string>();
-  for (const d of devices ?? []) {
-    deviceToUser.set(d.id as string, d.user_id as string);
-  }
-
-  const ownerIds = Array.from(new Set(deviceToUser.values()));
-  const { data: ownerUsers } =
-    ownerIds.length > 0
-      ? await admin
-          .from("users")
-          .select("id, display_name, email, org_id")
-          .in("id", ownerIds)
-      : { data: [] as UserLookup[] };
-
-  // Which owner IDs should surface by name to this viewer?
-  //   - Manager: every owner in their org
-  //   - Member:  only themselves; anything else collapses into Unassigned
-  const visibleOwnerIds = new Set<string>(
-    user.role === "manager"
-      ? (ownerUsers ?? [])
-          .filter((u) => u.org_id === user.org_id)
-          .map((u) => u.id)
-      : [user.id]
-  );
-
-  const userMeta = new Map<string, string>();
-  for (const u of ownerUsers ?? []) {
-    userMeta.set(u.id, u.display_name || u.email || u.id.slice(0, 8));
-  }
-
-  type Bucket = {
-    id: string;
-    name: string;
-    cost_cents: number;
-    input_tokens: number;
-    output_tokens: number;
-  };
-  const byUser = new Map<string, Bucket>();
-  for (const r of deviceCostRows) {
-    const ownerId = deviceToUser.get(r.device_id);
-    const bucketId =
-      ownerId && visibleOwnerIds.has(ownerId) ? ownerId : UNASSIGNED_USER_ID;
-    const cost = Number(r.cost_cents);
-    const inTok = Number(r.input_tokens ?? 0);
-    const outTok = Number(r.output_tokens ?? 0);
-    const existing = byUser.get(bucketId);
-    if (existing) {
-      existing.cost_cents += cost;
-      existing.input_tokens += inTok;
-      existing.output_tokens += outTok;
-    } else {
-      byUser.set(bucketId, {
-        id: bucketId,
-        name:
-          bucketId === UNASSIGNED_USER_ID
-            ? "Unassigned"
-            : (userMeta.get(bucketId) ?? bucketId.slice(0, 8)),
-        cost_cents: cost,
-        input_tokens: inTok,
-        output_tokens: outTok,
-      });
-    }
-  }
-
-  return Array.from(byUser.values())
-    .filter((u) => u.cost_cents > 0 || u.input_tokens + u.output_tokens > 0)
-    .sort((a, b) => {
-      // Keep "Unassigned" at the end regardless of its magnitude.
-      if (a.id === UNASSIGNED_USER_ID) return 1;
-      if (b.id === UNASSIGNED_USER_ID) return -1;
-      return b.cost_cents - a.cost_cents;
-    });
-}
-
-interface DeviceCostRow {
-  device_id: string;
-  cost_cents: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-export interface TeamActivityDay {
-  bucket_day: string;
-  active_members: number;
-  cost_cents: number;
-  input_tokens: number;
-  output_tokens: number;
-}
-
-/**
- * Daily series of distinct active members + total cost for the Team page (#127).
- * Active = any device the user owns wrote a rollup row for that bucket.
- * Manager sees the full org; member sees own devices only (ADR-0083 §6) — i.e.
- * a member's chart will read 0 or 1 every day, which is the same self-only
- * answer the rest of the page already gives them.
- *
- * Days with no rollup activity simply don't appear in the result; the chart
- * components decide whether to interpolate or render a gap.
- */
-export async function getTeamActivityByDay(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-): Promise<TeamActivityDay[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_team_activity_by_day", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as TeamActivityRow[])
-    .map((r) => ({
-      bucket_day: r.bucket_day,
-      active_members: Number(r.active_members),
-      cost_cents: Number(r.cost_cents),
-      input_tokens: Number(r.input_tokens ?? 0),
-      output_tokens: Number(r.output_tokens ?? 0),
-    }))
-    .sort((a, b) => a.bucket_day.localeCompare(b.bucket_day));
-}
-
-interface TeamActivityRow {
-  bucket_day: string;
-  active_members: number | string;
-  cost_cents: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-export interface DeviceActivityDay {
-  bucket_day: string;
-  active_devices: number;
-  cost_cents: number;
-  input_tokens: number;
-  output_tokens: number;
-}
-
-/**
- * Daily series of distinct active devices + total cost for the Devices page
- * (#145). Active = the device wrote any rollup row for that bucket. Manager
- * sees the full org; member sees own devices only (ADR-0083 §6). When the
- * manager's `UserFilter` is engaged the series is narrowed to that teammate's
- * devices so it stays consistent with the per-device bar chart on the same
- * page.
- *
- * Days with no rollup activity simply don't appear in the result; the chart
- * components decide whether to interpolate or render a gap.
- */
-export async function getDeviceActivityByDay(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-): Promise<DeviceActivityDay[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_device_activity_by_day", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as DeviceActivityRow[])
-    .map((r) => ({
-      bucket_day: r.bucket_day,
-      active_devices: Number(r.active_devices),
-      cost_cents: Number(r.cost_cents),
-      input_tokens: Number(r.input_tokens ?? 0),
-      output_tokens: Number(r.output_tokens ?? 0),
-    }))
-    .sort((a, b) => a.bucket_day.localeCompare(b.bucket_day));
-}
-
-interface DeviceActivityRow {
-  bucket_day: string;
-  active_devices: number | string;
-  cost_cents: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-export interface ModelActivityDay {
-  bucket_day: string;
-  active_models: number;
-  cost_cents: number;
-  input_tokens: number;
-  output_tokens: number;
-}
-
-/**
- * Daily series of distinct active models + total cost for the Models page
- * (#147). Active = the `(provider, model)` pair has any rollup row in the
- * bucket. Manager sees the full org; member sees own devices only
- * (ADR-0083 §6). When the manager's `UserFilter` is engaged the series is
- * narrowed to that teammate's devices so it stays consistent with the
- * per-model bar chart on the same page.
- *
- * Days with no rollup activity simply don't appear in the result; the chart
- * components decide whether to interpolate or render a gap.
- */
-export async function getModelActivityByDay(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-): Promise<ModelActivityDay[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_model_activity_by_day", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as ModelActivityRow[])
-    .map((r) => ({
-      bucket_day: r.bucket_day,
-      active_models: Number(r.active_models),
-      cost_cents: Number(r.cost_cents),
-      input_tokens: Number(r.input_tokens ?? 0),
-      output_tokens: Number(r.output_tokens ?? 0),
-    }))
-    .sort((a, b) => a.bucket_day.localeCompare(b.bucket_day));
-}
-
-interface ModelActivityRow {
-  bucket_day: string;
-  active_models: number | string;
-  cost_cents: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-interface UserLookup {
-  id: string;
-  display_name: string | null;
-  email: string | null;
-  org_id: string | null;
-}
-
-export interface DeviceCost {
-  id: string;
-  label: string | null;
-  owner_name: string | null;
-  last_seen: string | null;
-  cost_cents: number;
-  input_tokens: number;
-  output_tokens: number;
-}
-
-/**
- * Get cost breakdown by device.
- * Manager sees every device in the org; member sees only their own (ADR-0083 §6).
- *
- * Reuses `getVisibleDeviceIds` so the device set — and therefore the rollup
- * sum — matches Overview / Team for the same (user, range). For any visible
- * user, the Overview total equals the sum of `cost_cents` returned here.
- *
- * `owner_name` is populated in the manager view so the table can disambiguate
- * two laptops labelled `"laptop"` sitting under different owners. For members
- * it stays `null` — the member already knows every row is theirs and the
- * extra column would just be noise.
- */
-export async function getCostByDevice(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-): Promise<DeviceCost[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  // Aggregate rollups server-side (#92). The cap-truncation that #15 / #90
-  // chased lives entirely on the rollup-row pull, so removing that pull is
-  // the fix; the secondary metadata reads below are bounded by org device
-  // count and never exceed the 1k PostgREST default.
-  const { data: rows, error } = await admin.rpc("dashboard_cost_by_device", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-  const rollups = (rows ?? []) as DeviceCostRow[];
-
-  const { data: devices } = await admin
-    .from("devices")
-    .select("id, label, user_id, last_seen")
-    .in("id", deviceIds);
-
-  const deviceMeta = new Map<
-    string,
-    { label: string | null; user_id: string; last_seen: string | null }
-  >();
-  for (const d of devices ?? []) {
-    deviceMeta.set(d.id as string, {
-      label: (d.label as string | null) ?? null,
-      user_id: d.user_id as string,
-      last_seen: (d.last_seen as string | null) ?? null,
-    });
-  }
-
-  let ownerLookup = new Map<string, string>();
-  if (user.role === "manager") {
-    const ownerIds = Array.from(
-      new Set(Array.from(deviceMeta.values()).map((d) => d.user_id))
-    );
-    if (ownerIds.length > 0) {
-      const { data: owners } = await admin
-        .from("users")
-        .select("id, display_name, email")
-        .in("id", ownerIds);
-      ownerLookup = new Map(
-        (owners ?? []).map((u) => [
-          u.id as string,
-          (u.display_name as string | null) ||
-            (u.email as string | null) ||
-            (u.id as string).slice(0, 8),
-        ])
-      );
-    }
-  }
-
-  // The RPC already aggregates by device_id, so each row is a (device, sum)
-  // tuple — no further reduction needed.
-  const totalsByDevice = new Map<
-    string,
-    { cost_cents: number; input_tokens: number; output_tokens: number }
-  >();
-  for (const r of rollups) {
-    totalsByDevice.set(r.device_id, {
-      cost_cents: Number(r.cost_cents),
-      input_tokens: Number(r.input_tokens ?? 0),
-      output_tokens: Number(r.output_tokens ?? 0),
-    });
-  }
-
-  // Surface every visible device — including zero-cost ones — so a brand-new
-  // daemon shows up the moment it registers, even before it has pushed a
-  // rollup. That matters most during the "linked, waiting for first sync"
-  // gap covered by FirstSyncInProgressBanner on Overview.
-  const result: DeviceCost[] = [];
-  for (const [id, meta] of deviceMeta) {
-    const totals = totalsByDevice.get(id);
-    result.push({
-      id,
-      label: meta.label,
-      owner_name:
-        user.role === "manager"
-          ? (ownerLookup.get(meta.user_id) ?? null)
-          : null,
-      last_seen: meta.last_seen,
-      cost_cents: totals?.cost_cents ?? 0,
-      input_tokens: totals?.input_tokens ?? 0,
-      output_tokens: totals?.output_tokens ?? 0,
-    });
-  }
-
-  return result.sort((a, b) => b.cost_cents - a.cost_cents);
-}
-
-/**
- * Get cost breakdown by model.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- */
-export async function getCostByModel(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-) {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_cost_by_model", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as ModelCostRow[])
-    .map((r) => {
-      const effective = Number(r.cost_cents);
-      return {
-        provider: r.provider,
-        model: r.model,
-        cost_cents: effective,
-        // Pre-022 the RPC didn't surface `_ingested`; fall back to effective so
-        // the "List / Effective" tooltip stays hidden on deployments running
-        // ahead of migration 022 rather than rendering "$X / $X" everywhere.
-        cost_cents_ingested: Number(r.cost_cents_ingested ?? effective),
-        input_tokens: Number(r.input_tokens ?? 0),
-        output_tokens: Number(r.output_tokens ?? 0),
-      };
-    })
-    .filter((m) => m.cost_cents > 0 || m.input_tokens + m.output_tokens > 0)
-    .sort((a, b) => b.cost_cents - a.cost_cents);
-}
-
-interface ModelCostRow {
-  provider: string;
-  model: string;
-  cost_cents: number | string;
-  cost_cents_ingested?: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-/**
- * Get cost breakdown by repo.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- */
-export async function getCostByRepo(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-) {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_cost_by_repo", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as RepoCostRow[])
-    .map((r) => {
-      const effective = Number(r.cost_cents);
-      return {
-        repo_id: r.repo_id,
-        cost_cents: effective,
-        cost_cents_ingested: Number(r.cost_cents_ingested ?? effective),
-        input_tokens: Number(r.input_tokens ?? 0),
-        output_tokens: Number(r.output_tokens ?? 0),
-      };
-    })
-    .filter((r) => r.cost_cents > 0 || r.input_tokens + r.output_tokens > 0)
-    .sort((a, b) => b.cost_cents - a.cost_cents);
-}
-
-interface RepoCostRow {
-  repo_id: string;
-  cost_cents: number | string;
-  cost_cents_ingested?: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-/**
- * Get cost breakdown by branch.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- */
-export async function getCostByBranch(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-) {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_cost_by_branch", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as BranchCostRow[])
-    .map((r) => {
-      const effective = Number(r.cost_cents);
-      return {
-        repo_id: r.repo_id,
-        git_branch: r.git_branch,
-        cost_cents: effective,
-        cost_cents_ingested: Number(r.cost_cents_ingested ?? effective),
-        input_tokens: Number(r.input_tokens ?? 0),
-        output_tokens: Number(r.output_tokens ?? 0),
-      };
-    })
-    .filter((b) => b.cost_cents > 0 || b.input_tokens + b.output_tokens > 0)
-    .sort((a, b) => b.cost_cents - a.cost_cents);
-}
-
-interface BranchCostRow {
-  repo_id: string;
-  git_branch: string;
-  cost_cents: number | string;
-  cost_cents_ingested?: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-/**
- * Get cost breakdown by ticket.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- */
-export async function getCostByTicket(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-) {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_cost_by_ticket", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as TicketCostRow[])
-    .map((r) => {
-      const effective = Number(r.cost_cents);
-      return {
-        ticket: r.ticket,
-        cost_cents: effective,
-        cost_cents_ingested: Number(r.cost_cents_ingested ?? effective),
-        input_tokens: Number(r.input_tokens ?? 0),
-        output_tokens: Number(r.output_tokens ?? 0),
-      };
-    })
-    .filter((t) => t.cost_cents > 0 || t.input_tokens + t.output_tokens > 0)
-    .sort((a, b) => b.cost_cents - a.cost_cents);
-}
-
-interface TicketCostRow {
-  ticket: string;
-  cost_cents: number | string;
-  cost_cents_ingested?: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-export interface SurfaceCost {
-  surface: string;
-  cost_cents: number;
-  input_tokens: number;
-  output_tokens: number;
-}
-
-/**
- * Cost share by surface for the "Spend by Surface" card on the dashboard
- * (#187 part 2). Manager sees full org; member sees own devices only
- * (ADR-0083 §6). `options.scopedUserId` further narrows a manager view to a
- * single teammate; `options.surfaces` (from the chip) further narrows the
- * surfaces aggregated over — useful for the "show only `vscode` and
- * `cursor`" comparison.
- *
- * Returned rows include `unknown` whenever a device has rollups from a
- * pre-bump daemon, matching the issue acceptance ("rows from pre-bump
- * daemons display as `unknown` and remain in the all-surfaces aggregation").
- * The empty-state for single-surface orgs (one bar, full width) is the
- * caller's call — we always return whatever is present in the data.
- */
-export async function getCostBySurface(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions
-): Promise<SurfaceCost[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_cost_by_surface", {
-    p_device_ids: deviceIds,
-    p_bucket_from: range.bucketFrom,
-    p_bucket_to: range.bucketTo,
-    p_surfaces: normalizeSurfaces(options?.surfaces),
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as SurfaceCostRow[])
-    .map((r) => ({
-      surface: r.surface,
-      cost_cents: Number(r.cost_cents),
-      input_tokens: Number(r.input_tokens ?? 0),
-      output_tokens: Number(r.output_tokens ?? 0),
-    }))
-    .filter((s) => s.cost_cents > 0 || s.input_tokens + s.output_tokens > 0)
-    .sort((a, b) => b.cost_cents - a.cost_cents);
-}
-
-interface SurfaceCostRow {
-  surface: string;
-  cost_cents: number | string;
-  input_tokens?: number | string;
-  output_tokens?: number | string;
-}
-
-/**
- * Distinct surfaces with at least one rollup row visible to the viewer.
- * Powers the `<SurfaceFilter>` chip's options (#187 part 1) — drawn from
- * data so the day a JetBrains daemon first syncs the chip picks up
- * `jetbrains` automatically; no hardcoded enum to keep in sync with core.
- *
- * Deliberately not range-scoped: a surface that appeared in the org's
- * history but not in the current period still belongs in the chip so the
- * "filter to JetBrains" path remains usable after the team migrates off it.
- * `unknown` is included when present so a manager investigating "rows
- * without a surface tag" can still drill in.
- */
-export async function getKnownSurfaces(
-  user: BudiUser,
-  options?: ScopeOptions
-): Promise<string[]> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return [];
-
-  const { data, error } = await admin.rpc("dashboard_known_surfaces", {
-    p_device_ids: deviceIds,
-  });
-  if (error) throw error;
-
-  return ((data ?? []) as { surface: string }[])
-    .map((r) => r.surface)
-    .filter((s) => typeof s === "string" && s.length > 0);
-}
-
-/**
- * Cursor for `getSessions` pagination — encodes the row at the boundary of the
- * current page. The composite `(started_at, session_id)` shape gives a stable
- * walk even when two sessions share a `started_at` (rare but real once cursor
- * sync re-emits a batch with the same instant): the SQL filter
- * `(started_at, session_id) < cursor` keeps the strict ordering and never
- * skips or duplicates a tied row.
- *
- * The dashboard URL serializes this as `?cursor=<base64url(JSON)>` so a
- * session_id containing punctuation never collides with a delimiter (#85).
- */
-export interface SessionsCursor {
-  startedAt: string;
-  sessionId: string;
-}
-
-/** Default page size for the Sessions table — matches the UI's pager. */
-export const SESSIONS_PAGE_SIZE = 50;
-
-/**
- * Get a single page of sessions ordered by `(started_at desc, session_id desc)`.
- * Manager sees full org; member sees own devices only (ADR-0083 §6).
- * `options.scopedUserId` further narrows a manager view to a single teammate.
- *
- * Pagination is cursor-based on `(started_at, session_id)`. Sessions are
- * immutable once written so the cursor is stable across reloads — no offset
- * skew under concurrent writes, no expensive count query required to know if
- * another page exists. We fetch `pageSize + 1` rows so `hasMore` falls out of
- * the result-set size; the extra row is dropped before returning.
- *
- * History: prior to #85 this returned the most recent 100 rows with no
- * pagination, silently truncating the visible Sessions history to whatever
- * fit in those 100 rows (~9 days for a high-volume org). The
- * `Recent Sessions (100+)` title was the only hint that anything older
- * existed. Cursor pagination replaces both.
- */
-export async function getSessions(
-  user: BudiUser,
-  range: DateRange,
-  options?: ScopeOptions,
-  pagination?: { pageSize?: number; cursor?: SessionsCursor | null }
-): Promise<{ rows: SessionRow[]; nextCursor: SessionsCursor | null }> {
-  const admin = createAdminClient();
-  const deviceIds = await getVisibleDeviceIds(admin, user, options);
-  if (deviceIds.length === 0) return { rows: [], nextCursor: null };
-
-  const pageSize = pagination?.pageSize ?? SESSIONS_PAGE_SIZE;
-  const cursor = pagination?.cursor ?? null;
-
-  const surfaces = normalizeSurfaces(options?.surfaces);
-  let query = admin
-    .from("session_summaries")
-    .select("*")
-    .in("device_id", deviceIds)
-    .gte("started_at", range.startedAtFrom)
-    .lte("started_at", range.startedAtTo)
-    .order("started_at", { ascending: false })
-    // Tie-breaker: without a secondary sort key, two rows with the same
-    // `started_at` could appear on either side of a cursor boundary across
-    // requests, causing rows to skip or duplicate as the user paginates.
-    .order("session_id", { ascending: false })
-    .limit(pageSize + 1);
-
-  if (surfaces) query = query.in("surface", surfaces);
-
-  if (cursor) {
-    // Composite tuple compare: (started_at, session_id) < cursor.
-    // PostgREST has no native row-constructor compare, so we expand to the
-    // logically-equivalent disjunction.
-    //
-    // Both values are wrapped in PostgREST's quoted-literal syntax — without
-    // the quoting, a `session_id` containing filter-tree metacharacters (`,`,
-    // `(`, `)`) would inject extra top-level conditions into the disjunction,
-    // breaking the cursor invariant. Decoding already rejects those shapes
-    // (#176), but we quote here too so a direct DAL caller can't bypass it.
-    const startedAt = postgrestQuoteLiteral(cursor.startedAt);
-    const sessionId = postgrestQuoteLiteral(cursor.sessionId);
-    query = query.or(
-      `started_at.lt.${startedAt},and(started_at.eq.${startedAt},session_id.lt.${sessionId})`
-    );
-  }
-
-  const { data } = await query;
-  const fetched = (data ?? []) as SessionRow[];
-  const hasMore = fetched.length > pageSize;
-  const trimmed = hasMore ? fetched.slice(0, pageSize) : fetched;
-  const rows = await attachOwners(admin, user, trimmed);
-  const tail = rows[rows.length - 1];
-  const nextCursor =
-    hasMore && tail
-      ? { startedAt: tail.started_at, sessionId: tail.session_id }
-      : null;
-
-  return { rows, nextCursor };
-}
-
-/**
- * Resolve owner display labels for a batch of session rows. Manager-only:
- * member viewers already know every row is theirs (#138), so we leave
- * `owner_name` null and skip the device→user→identity joins entirely.
- *
- * Falls back through `display_name → email → id-prefix` so a freshly-invited
- * teammate without a profile name still renders something a manager can match
- * to a person in the team list. Mirrors the lookup already proven out in
- * `getCostByDevice` so the two surfaces label the same teammate identically.
- */
-async function attachOwners(
-  admin: ReturnType<typeof createAdminClient>,
-  user: BudiUser,
-  rows: SessionRow[]
-): Promise<SessionRow[]> {
-  if (user.role !== "manager" || rows.length === 0) return rows;
-
-  const deviceIds = Array.from(new Set(rows.map((r) => r.device_id)));
-  const { data: devices } = await admin
-    .from("devices")
-    .select("id, user_id")
-    .in("id", deviceIds);
-  const deviceToUser = new Map<string, string>();
-  for (const d of devices ?? []) {
-    deviceToUser.set(d.id as string, d.user_id as string);
-  }
-
-  const ownerIds = Array.from(new Set(deviceToUser.values()));
-  if (ownerIds.length === 0) return rows;
-
-  const { data: owners } = await admin
-    .from("users")
-    .select("id, display_name, email")
-    .in("id", ownerIds);
-  const ownerLookup = new Map<string, string>(
-    (owners ?? []).map((u) => [
-      u.id as string,
-      (u.display_name as string | null) ||
-        (u.email as string | null) ||
-        (u.id as string).slice(0, 8),
-    ])
-  );
-
-  return rows.map((r) => {
-    const ownerId = deviceToUser.get(r.device_id);
-    return {
-      ...r,
-      owner_name: ownerId ? (ownerLookup.get(ownerId) ?? null) : null,
-    };
-  });
-}
-
-export interface SessionRow {
-  device_id: string;
-  session_id: string;
-  provider: string;
-  started_at: string;
-  ended_at: string | null;
-  duration_ms: number | null;
-  repo_id: string | null;
-  git_branch: string | null;
-  ticket: string | null;
-  message_count: number;
-  total_input_tokens: number | string;
-  total_output_tokens: number | string;
-  // Dual cost columns (#231): `_effective` is what the dashboard renders;
-  // `_ingested` is the daemon-uploaded value before any team-pricing recalc
-  // was applied. Until the recalc engine ships (#233) the two are equal; the
-  // "list-vs-effective delta + savings" widget (#235) will surface the gap.
-  total_cost_cents_effective: number | string;
-  total_cost_cents_ingested: number | string;
-  // Per-session main model (#140). NULL for rows ingested before the daemon
-  // started emitting `primary_model`, and for sessions with zero scored
-  // messages — render as em-dash in those cases.
-  main_model: string | null;
-  // Free-form session title (#255). Sources include the daemon's parsed
-  // `session_title` tag (siropkin/budi#779) — typically an IntelliJ project
-  // name (`Verkada-Web`) or session-type label (`chat-agent`). NULL for
-  // pre-8.5.0 rows and for surfaces that don't emit a title; render as a
-  // muted dash in those cases.
-  title: string | null;
-  // Resolved owner label for the device this session ran on (#138). Only
-  // populated for manager viewers; null for member viewers (every row is
-  // theirs) and for sessions whose device→user mapping cannot be resolved.
-  // Not a column on `session_summaries` — joined in by the DAL via
-  // `attachOwners`.
-  owner_name?: string | null;
-  // Surface dimension (#187). Pre-#187 rows backfill to the literal
-  // `'unknown'`, so the Sessions table column never has a null hole — the
-  // dashboard treats `unknown` as its own bucket and filters/displays
-  // accordingly.
-  surface: string;
-  // The schema also has `vital_*` columns (006_session_vitals.sql) but the
-  // daemon has never populated them, so the dashboard stopped reading them in
-  // #141. Reintroduce typed fields here once budi-core ships vitals on the
-  // ingest envelope.
-}
-
-/**
- * Fetch a single session by `(device_id, session_id)` for the session-detail
- * page (#99). Returns `null` when the session does not exist *or* when it
- * exists but is not visible to the viewer (manager: anywhere in the org;
- * member: only on a device they own — same scoping as `getSessions`). The
- * "not visible" → `null` branch deliberately collapses with "not found" so
- * the URL parameter cannot be used to probe whether a foreign-org session
- * exists.
- */
-export async function getSessionDetail(
-  user: BudiUser,
-  deviceId: string,
-  sessionId: string
-): Promise<SessionRow | null> {
-  const admin = createAdminClient();
-  const visibleDeviceIds = await getVisibleDeviceIds(admin, user);
-  if (!visibleDeviceIds.includes(deviceId)) return null;
-
-  const { data } = await admin
-    .from("session_summaries")
-    .select("*")
-    .eq("device_id", deviceId)
-    .eq("session_id", sessionId)
-    .maybeSingle();
-
-  const row = (data as SessionRow | null) ?? null;
-  if (!row) return null;
-  const [enriched] = await attachOwners(admin, user, [row]);
-  return enriched ?? row;
-}
-
-/**
- * Fetch a single session by `session_id` alone, scoped to the viewer's
- * visible devices. Powers the deep-link entry point at
- * `/dashboard/sessions/<id>` without `?device=` (#202): a manager pasting a
- * session URL from chat / a ticket should land on the page even though the
- * URL doesn't carry the device half of the composite PK.
- *
- * Returns `null` when no row matches OR when more than one device in the
- * viewer's scope happens to share the same `session_id` (extraordinarily
- * rare for UUID daemons, but the ambiguous case must collapse with
- * not-found rather than guess and silently render the wrong row).
- *
- * Existence is collapsed with visibility — same privacy contract as
- * `getSessionDetail`: a foreign-org session never reveals its existence
- * via response shape or timing differences. Callers should `notFound()`
- * on a `null` return.
- */
-export async function getSessionDetailBySessionId(
-  user: BudiUser,
-  sessionId: string
-): Promise<SessionRow | null> {
-  const admin = createAdminClient();
-  const visibleDeviceIds = await getVisibleDeviceIds(admin, user);
-  if (visibleDeviceIds.length === 0) return null;
-
-  const { data } = await admin
-    .from("session_summaries")
-    .select("*")
-    .in("device_id", visibleDeviceIds)
-    .eq("session_id", sessionId)
-    // Cap at 2 so the ambiguous-match branch can short-circuit without
-    // pulling an unbounded result set when the daemon-emitted session_id
-    // happens to be a non-UUID string that appears across many devices.
-    .limit(2);
-
-  const rows = (data ?? []) as SessionRow[];
-  if (rows.length !== 1) return null;
-  const [enriched] = await attachOwners(admin, user, rows);
-  return enriched ?? rows[0]!;
-}
-
-/**
- * Sync freshness snapshot for the viewer.
- *
- * Used by the dashboard header to render a "Last synced X ago" indicator and
- * to distinguish *not linked yet* from *linked, waiting for first sync* from
- * *stalled*. Always scoped to the **viewer's own devices**, regardless of
- * role — the badge answers "is *my* daemon healthy" and a manager whose own
- * daemon has been silent for hours should see a stale badge even if a
- * teammate's daemon synced 3 minutes ago (#74). The org-wide "who's stale?"
- * view lives on `/dashboard/devices`, which is the right surface for a
- * manager chasing down a teammate's broken daemon.
- *
- * - `deviceCount` is the number of daemons the viewer themselves has linked.
- *   Zero means the viewer's account exists on cloud but they haven't run
- *   `budi cloud init` yet — the "not linked yet" state. (The same field
- *   drives the `LinkDaemonBanner` on the overview, so a manager who hasn't
- *   linked their own daemon still gets the prompt even if teammates have.)
- * - `lastSeenAt` is the most recent `devices.last_seen` across the viewer's
- *   own devices. It advances on every successful ingest, even when the
- *   payload contains zero rollups, so it's the authoritative "is *my*
- *   daemon talking to us" signal.
- * - `lastRollupAt` is the most recent `daily_rollups.synced_at` across the
- *   viewer's own devices. If `deviceCount > 0` but `lastRollupAt` is null,
- *   the viewer is linked but hasn't pushed any usage rows yet — that's the
- *   "initial sync in progress / no data yet" state.
- * - `lastSessionAt` is the most recent `session_summaries.started_at` across
- *   the viewer's own devices. The two ingest streams (rollups and sessions)
- *   travel in the same envelope but are independent rows on the cloud side,
- *   so a daemon-side regression can drop sessions while still landing
- *   rollups. The header badge cross-checks the two watermarks so the
- *   divergence surfaces in the badge instead of silently emptying the
- *   Sessions page (#84).
- */
-export async function getSyncFreshness(user: BudiUser): Promise<{
-  deviceCount: number;
-  lastSeenAt: string | null;
-  lastRollupAt: string | null;
-  lastSessionAt: string | null;
-}> {
-  const admin = createAdminClient();
-  const { data: ownDevices } = await admin
-    .from("devices")
-    .select("id")
-    .eq("user_id", user.id);
-  const deviceIds = (ownDevices ?? []).map((d) => d.id as string);
-  if (deviceIds.length === 0) {
-    return {
-      deviceCount: 0,
-      lastSeenAt: null,
-      lastRollupAt: null,
-      lastSessionAt: null,
-    };
-  }
-
-  const { data: lastSeenRow } = await admin
-    .from("devices")
-    .select("last_seen")
-    .in("id", deviceIds)
-    .order("last_seen", { ascending: false })
-    .limit(1)
-    .single();
-
-  // `.maybeSingle()` because the "linked but no rollups yet" first-run state
-  // legitimately returns zero rows (covered by `FirstSyncInProgressBanner`).
-  // `.single()` would emit a PGRST116 / HTTP 406 row in production logs on
-  // every freshly-linked-no-data render even though the call site already
-  // tolerates `data: null` (see #22).
-  const { data: lastRollupRow } = await admin
-    .from("daily_rollups")
-    .select("synced_at")
-    .in("device_id", deviceIds)
-    .order("synced_at", { ascending: false })
-    .limit(1)
-    .maybeSingle();
-
-  // Probe `started_at` rather than `synced_at`: the user-facing Sessions
-  // page sorts by `started_at`, and a "session that never closes" daemon
-  // bug would keep upserting the same row's `synced_at` while no new
-  // sessions land, masking the divergence we want to detect.
-  const { data: lastSessionRow } = await admin
-    .from("session_summaries")
-    .select("started_at")
-    .in("device_id", deviceIds)
-    .order("started_at", { ascending: false })
-    .limit(1)
-    .maybeSingle();
-
-  return {
-    deviceCount: deviceIds.length,
-    lastSeenAt: (lastSeenRow?.last_seen as string | null) ?? null,
-    lastRollupAt: (lastRollupRow?.synced_at as string | null) ?? null,
-    lastSessionAt: (lastSessionRow?.started_at as string | null) ?? null,
-  };
-}
-
-/**
- * Whether the org has at least one `active` price list. Powers the
- * conditional rendering of the savings strip and the Effective/List toggle
- * on the Overview (#235): when no list is active, the daemon-uploaded cost
- * is the only number that exists, so there is nothing to compare against
- * and the UI stays clean. Until the upload UI ships (#232) this returns
- * `false` for every org — and the Overview keeps rendering exactly as it
- * does today, which is the acceptance criterion for migration 019.
- */
-export async function getOrgHasActivePriceList(
-  orgId: string
-): Promise<boolean> {
-  const admin = createAdminClient();
-  const { data, error } = await admin
-    .from("org_price_lists")
-    .select("id")
-    .eq("org_id", orgId)
-    .eq("status", "active")
-    .limit(1)
-    .maybeSingle();
-  if (error) return false;
-  return data !== null;
-}
-
-/**
- * Page of `recalculation_runs` rows for the Settings → Pricing audit-history
- * tab (#733). Each row is one invocation of the recalc engine (#728); the
- * rows are immutable. We support a simple offset window because the table is
- * small (one row per admin click) and the surface paginates 50 at a time.
- *
- * `status` narrows by `recalculation_runs.status`. Pass `null` to include
- * every run. The page also returns `total` so the UI can size the pager
- * without a second round-trip — `count: "exact"` is cheap on this table.
- */
-export interface RecalculationRunRow {
-  id: number;
-  startedAt: string;
-  finishedAt: string | null;
-  status: string;
-  scopeFromDate: string | null;
-  scopeToDate: string | null;
-  priceListIds: number[];
-  rowsProcessed: number | null;
-  rowsChanged: number | null;
-  beforeTotalCents: number | null;
-  afterTotalCents: number | null;
-  triggeredBy: string | null;
-}
-
-export async function getRecalculationRuns(
-  orgId: string,
-  options: { status: string | null; limit: number; offset: number }
-): Promise<{ rows: RecalculationRunRow[]; total: number }> {
-  const admin = createAdminClient();
-  let query = admin
-    .from("recalculation_runs")
-    .select(
-      "id, started_at, finished_at, status, scope_from_date, scope_to_date, price_list_ids, rows_processed, rows_changed, before_total_cents, after_total_cents, triggered_by",
-      { count: "exact" }
-    )
-    .eq("org_id", orgId)
-    .order("started_at", { ascending: false })
-    .range(options.offset, options.offset + options.limit - 1);
-  if (options.status) query = query.eq("status", options.status);
-
-  const { data, count, error } = await query;
-  if (error) throw error;
-
-  const rows: RecalculationRunRow[] = (data ?? []).map((r) => ({
-    id: r.id as number,
-    startedAt: r.started_at as string,
-    finishedAt: (r.finished_at as string | null) ?? null,
-    status: (r.status as string) ?? "",
-    scopeFromDate: (r.scope_from_date as string | null) ?? null,
-    scopeToDate: (r.scope_to_date as string | null) ?? null,
-    priceListIds: ((r.price_list_ids as number[] | null) ?? []).map(Number),
-    rowsProcessed: r.rows_processed == null ? null : Number(r.rows_processed),
-    rowsChanged: r.rows_changed == null ? null : Number(r.rows_changed),
-    beforeTotalCents:
-      r.before_total_cents == null ? null : Number(r.before_total_cents),
-    afterTotalCents:
-      r.after_total_cents == null ? null : Number(r.after_total_cents),
-    triggeredBy: (r.triggered_by as string | null) ?? null,
-  }));
-  return { rows, total: count ?? 0 };
-}
-
-/**
- * Get org members list.
- */
-export async function getOrgMembers(orgId: string) {
-  const admin = createAdminClient();
-  const { data } = await admin
-    .from("users")
-    .select("id, display_name, email, role, created_at")
-    .eq("org_id", orgId)
-    .order("created_at");
-
-  return data ?? [];
-}
-
-// --- Helpers ---
-
-/**
- * Wrap a value in PostgREST's quoted-literal syntax (`"..."`). Escapes
- * embedded `\` and `"` per PostgREST's docs so the filter tree always parses
- * as a single leaf — never as a stray delimiter that opens a new branch in
- * the parent expression. Used by `getSessions` when expanding a composite
- * `(started_at, session_id) < cursor` compare into a `.or()` disjunction.
- */
-function postgrestQuoteLiteral(value: string): string {
-  const escaped = value.replace(/\\/g, "\\\\").replace(/"/g, '\\"');
-  return `"${escaped}"`;
-}
-
-/**
- * Get device IDs visible to the current user.
- * Per ADR-0083 §6:
- *   - Manager: sees all devices in the org
- *   - Member: sees only their own devices
- *
- * `options.scopedUserId` (manager-only, #80) narrows the result further to a
- * single teammate's devices. If the id is missing, unknown, or belongs to
- * another org we silently fall back to the org-wide set rather than 4xxing,
- * so the URL parameter cannot be used to probe other orgs' user ids. Members
- * already collapse to themselves and ignore the option entirely.
- */
-async function getVisibleDeviceIds(
-  admin: ReturnType<typeof createAdminClient>,
-  user: BudiUser,
-  options?: ScopeOptions
-): Promise<string[]> {
-  if (user.role === "manager") {
-    return getOrgDeviceIds(admin, user.org_id!, options?.scopedUserId ?? null);
-  }
-  // Member: own devices only — `scopedUserId` is intentionally ignored.
-  const { data: devices } = await admin
-    .from("devices")
-    .select("id")
-    .eq("user_id", user.id);
-  return (devices ?? []).map((d) => d.id);
-}
-
-async function getOrgDeviceIds(
-  admin: ReturnType<typeof createAdminClient>,
-  orgId: string,
-  scopedUserId: string | null
-): Promise<string[]> {
-  const { data: users } = await admin
-    .from("users")
-    .select("id")
-    .eq("org_id", orgId);
-
-  if (!users?.length) return [];
-
-  const orgUserIds = users.map((u) => u.id as string);
-  // Narrow to a single teammate when the manager picked one — but only if
-  // they're actually in the manager's org. Anything else collapses back to
-  // org-wide so an out-of-org id can't leak the existence of another org.
-  const userIds =
-    scopedUserId && orgUserIds.includes(scopedUserId)
-      ? [scopedUserId]
-      : orgUserIds;
-
-  const { data: devices } = await admin
-    .from("devices")
-    .select("id")
-    .in("user_id", userIds);
-
-  return (devices ?? []).map((d) => d.id);
-}
